+++ src/main/kotlin/org/jetbrains/research/testgenie/tools/evosuite/ResultWatcher.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.testgenie.evosuite
+package org.jetbrains.research.testgenie.tools.evosuite
 
 import com.google.gson.Gson
 import com.google.gson.stream.JsonReader
