+++ src/main/kotlin/org/jetbrains/research/testgenie/tools/evosuite/generation/EvoSuiteErrorManager.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.testgenie.evosuite.generation
+package org.jetbrains.research.testgenie.tools.evosuite.generation
 
 import com.intellij.execution.process.OSProcessHandler
 import com.intellij.notification.NotificationGroupManager
