+++ src/main/kotlin/org/jetbrains/research/testgenie/tools/evosuite/validation/TestCaseEditor.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.research.testgenie.evosuite.validation
+package org.jetbrains.research.testgenie.tools.evosuite.validation
 
 import com.github.javaparser.JavaParser
 import com.github.javaparser.StaticJavaParser
